agents/sub_agents/service_desk_agent/requirements.txt
boto3
botocore
langgraph
langchain-aws
langchain-core
langchain-community
bedrock-agentcore
bedrock-agentcore-starter-toolkit